fix: Use correct data type
...to prevent overflowing values

+++ Traperto.ClamAvRest.Main/Startup.cs
@@ -54,7 +54,7 @@ namespace VirusScannerService
 
                 if (maxStreamSize != null)
                 {
-                    clamClient.MaxStreamSize = int.Parse(maxStreamSize); 
+                    clamClient.MaxStreamSize = long.Parse(maxStreamSize); 
                 }
             
                 if (maxChunkSize != null)
